pyproject.toml
[tool.poetry]
name = "langgrapg-vineet"
version = "0.1.0"
description = ""
authors = ["srush1608 <srushti.talwekar@coditas.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.8"
python-dotenv = "^1.0.1"
langchain-groq = "^0.2.1"
flask = "^3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
